:wrench: Add a target to run dev attached
This patch adds a small target to the Makefile that is identical to
`run-dev` except for the fact that it runs all the dockers attached.
This is useful in development if you wish to know everything that is
happening at the same time in the stack without having to manually run
`docker compose logs -f` yourself.

Signed-off-by: Amélie "Lymkwi" González <lymkwi@vulpinecitrus.info>