⚠ Fix compose nginx port (it should be 80)
The `NGINX_PORT` in all of our docker compose files should always remain
at `80`, because it describes the internal port (and we do not care
about changing it). A mistake overlooked in the previous change to the
nginx port with environment variables actually changed them so the
internal port of the nginx docker was the same as the external one… but
the redirection wasn't changing.